app/requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
marshmallow==3.5.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
pynsq==0.8.3
pytz==2019.3
redis==3.4.1
requests==2.23.0
SQLAlchemy==1.3.16
tornado==4.5.3
urllib3==1.25.8
vine==1.3.0
Werkzeug==1.0.1
WTForms==2.2.1
zipp==3.1.0
